GPT-OSS 120B fused decode: drop dead workaround code
Cleanup pass over the GPT-OSS bringup commits. All deletions, no
behavior change for the GPT-OSS e2e path. PCC=0.99 with trace.

- Revert AssignOp NonCacheableTrait. Added speculatively in
  7f06ecaa to keep an SRC re-zero assign out of const-eval, but
  the SRC use was reverted in d9679a70 (PCC regression). The
  three pre-existing AssignOp callers (collective_permute,
  collective_broadcast, all_to_all) never needed it.

- Drop dead helpers in TTNNDecomposeLayouts (canTilizeShardedOnDevice;
  no callers) and runtime reshape (selectReshapeMapMode /
  isNonTileAligned, immediately overridden by `if (true) mapMode =
  CACHE`); collapse to a direct CACHE constant with the metal#40612
  workaround comment.

- Drop runtime debug knobs in all_to_all_dispatch_metadata:
  TT_A2A_DISPATCH_METADATA_SHARDED_TOPO and TT_A2A_DISPATCH_ALGO env
  vars (both default to the only shipping configuration), plus the
  read-and-(void)-cast block for the deprecated drain_core flatbuffer
  field.

- Drop dead static_cast<int64_t>(getNumDevices()) etc. in
  MoeGPTDecodeOp::verify (I64Attr getters already return int64_t).

- Drop two stray blank lines (TTNNPipelines.cpp:424,
  topk_router_gpt.cpp:26).

7 files, +14 / -122.

+++ lib/Dialect/TTIR/IR/TTIROps.cpp
@@ -6005,11 +6005,11 @@ verifyReduceOp(llvm::function_ref<mlir::InFlightDiagnostic()> emitOpError,
   RankedTensorType fusedW0W1Type = getFusedW0W1().getType();
   RankedTensorType fusedW2Type = getFusedW2().getType();
   RankedTensorType resultType = getResult().getType();
-  int64_t numDevices = static_cast<int64_t>(getNumDevices());
-  int64_t clusterAxis = static_cast<int64_t>(getClusterAxis());
-  int64_t numExperts = static_cast<int64_t>(getNumExperts());
-  int64_t numExpertsPerTok = static_cast<int64_t>(getNumExpertsPerTok());
-  int64_t intermediateSize = static_cast<int64_t>(getIntermediateSize());
+  int64_t numDevices = getNumDevices();
+  int64_t clusterAxis = getClusterAxis();
+  int64_t numExperts = getNumExperts();
+  int64_t numExpertsPerTok = getNumExpertsPerTok();
+  int64_t intermediateSize = getIntermediateSize();
 
   // Activations / routing tensors follow the 4D layout used by the surrounding
   // MoE decode custom calls (tt.all_to_all_dispatch_metadata, tt.moe_gpt,
